requirements.txt
fuzzywuzzy == 0.18.0
moviepy == 1.0.3
numpy == 1.26.4
python-Levenshtein == 0.25.1
PyYAML == 6.0.2
SpeechRecognition == 3.10.4

# Dev requirements
autopep8  ==  2.3.1
flake8 == 7.1.1
jedi  ==  0.19.1
